The Scientist's Machine
Niharika is a scientist and she has designed a robo . The function of Robo is just to walk forward and backward continuously. One day Niharika went out of the lab and one of kid has taken the robo and started observing how it is walking. But there is ditch on forward end and backward which was not observed by the kid. So now we have to check whether the robo falls in ditch or not. If it falls whether it is forward ditch or backward ditch and time to fall in ditch